web/requirements.txt
flask>=3.0.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
numpy>=1.24.0
werkzeug>=3.0.0
